refactor: 라우터 상수 위치 변경
- config 폴더 생성에서 관리하도록 함(환경 설정 관련 상수 폴더)
